test: move testes de admin para a pasta só de teste
